+++ AnimeGen/Views/ViewController.swift
@@ -8,10 +8,11 @@
 import UIKit
 import Photos
 import Kingfisher
+import SafariServices
 
 enum ImageSource: String {
-    case picRe = "picRe"
     case waifuIm = "waifuIm"
+    case picRe = "picRe"
 }
 
 class ViewController: UIViewController {
@@ -19,14 +20,14 @@ class ViewController: UIViewController {
     @IBOutlet weak var refreshButton: UIButton!
     @IBOutlet weak var heartButton: UIButton!
     @IBOutlet weak var rewindButton: UIButton!
-    @IBOutlet weak var settingsButton: UIButton!
+    @IBOutlet weak var safariButton: UIButton!
     @IBOutlet weak var sourceButton: UIButton!
     
-    var currentSource: ImageSource = .picRe
+    var currentSource: ImageSource = .waifuIm
     var imageHistory: [URL] = []
     var currentImageIndex: Int = -1
     
-    private lazy var activityIndicator: UIActivityIndicatorView = {
+    lazy var activityIndicator: UIActivityIndicatorView = {
         let indicator = UIActivityIndicatorView(style: .medium)
         indicator.translatesAutoresizingMaskIntoConstraints = false
         indicator.hidesWhenStopped = true
@@ -57,13 +58,6 @@ class ViewController: UIViewController {
     }
     
     func createSourceMenu() -> UIMenu {
-        let picReAction = UIAction(title: "pic.re", image: nil, state: (currentSource == .picRe) ? .on : .off) { [weak self] (action) in
-            self?.currentSource = .picRe
-            self?.saveSelectedSource()
-            self?.loadNewImage()
-            self?.setupSourceButtonMenu()
-        }
-        
         let waifuImAction = UIAction(title: "waifu.im", image: nil, state: (currentSource == .waifuIm) ? .on : .off) { [weak self] (action) in
             self?.currentSource = .waifuIm
             self?.saveSelectedSource()
@@ -71,110 +65,25 @@ class ViewController: UIViewController {
             self?.setupSourceButtonMenu()
         }
         
-        return UIMenu(title: "Select Source", children: [picReAction, waifuImAction])
+        let picReAction = UIAction(title: "pic.re", image: nil, state: (currentSource == .picRe) ? .on : .off) { [weak self] (action) in
+            self?.currentSource = .picRe
+            self?.saveSelectedSource()
+            self?.loadNewImage()
+            self?.setupSourceButtonMenu()
+        }
+        
+        return UIMenu(title: "Select Source", children: [waifuImAction, picReAction])
     }
     
     func loadNewImage() {
         activityIndicator.startAnimating()
         
         switch currentSource {
-        case .picRe:
-            fetchImageFromPicRe()
         case .waifuIm:
             fetchImageFromWaifuIm()
+        case .picRe:
+            fetchImageFromPicRe()
         }
-    }
-    
-    func fetchImageFromPicRe() {
-        guard let url = URL(string: "https://pic.re/image") else {
-            print("Invalid URL")
-            return
-        }
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        
-        let task = URLSession.custom.dataTask(with: request) { [weak self] (data, response, error) in
-            guard let self = self else { return }
-            
-            if let error = error {
-                print("Error fetching image from pic.re: \(error)")
-                self.showErrorAlert(message: "Failed to load image from pic.re")
-                self.activityIndicator.stopAnimating()
-                return
-            }
-            
-            guard let data = data else {
-                print("No data received from pic.re")
-                self.showErrorAlert(message: "No data received from pic.re")
-                self.activityIndicator.stopAnimating()
-                return
-            }
-            
-            do {
-                if let json = try JSONSerialization.jsonObject(with: data, options: .allowFragments) as? [String: Any],
-                   let fileURLString = json["file_url"] as? String,
-                   let fileURL = URL(string: "https://" + fileURLString) {
-                    DispatchQueue.main.async {
-                        self.loadImage(from: fileURL)
-                    }
-                } else {
-                    print("Error parsing pic.re JSON or 'file_url' not found")
-                    self.showErrorAlert(message: "Error parsing pic.re JSON or 'file_url' not found")
-                    self.activityIndicator.stopAnimating()
-                }
-            } catch {
-                print("Error decoding pic.re JSON: \(error)")
-                self.showErrorAlert(message: "Error decoding pic.re JSON")
-                self.activityIndicator.stopAnimating()
-            }
-        }
-        
-        task.resume()
-    }
-    
-    func fetchImageFromWaifuIm() {
-        let url = URL(string: "https://api.waifu.im/search?is_nsfw=false")!
-        
-        let task = URLSession.custom.dataTask(with: url) { [weak self] (data, response, error) in
-            guard let self = self else { return }
-            
-            if let error = error {
-                print("Error fetching waifu.im image: \(error)")
-                self.showErrorAlert(message: "Failed to load image from waifu.im")
-                self.activityIndicator.stopAnimating()
-                return
-            }
-            
-            guard let data = data else {
-                print("No data received from waifu.im")
-                self.showErrorAlert(message: "No data received from waifu.im")
-                self.activityIndicator.stopAnimating()
-                return
-            }
-            
-            do {
-                if let json = try JSONSerialization.jsonObject(with: data, options: []) as? [String: Any],
-                   let images = json["images"] as? [[String: Any]],
-                   let firstImage = images.first,
-                   let imageUrlString = firstImage["url"] as? String,
-                   let imageUrl = URL(string: imageUrlString) {
-                    DispatchQueue.main.async {
-                        self.loadImage(from: imageUrl)
-                    }
-                } else {
-                    print("Error parsing waifu.im JSON")
-                    self.showErrorAlert(message: "Error parsing waifu.im JSON")
-                    self.activityIndicator.stopAnimating()
-                }
-            } catch {
-                print("Error decoding waifu.im JSON: \(error)")
-                self.showErrorAlert(message: "Error decoding waifu.im JSON")
-                self.activityIndicator.stopAnimating()
-            }
-        }
-        
-        task.resume()
     }
     
     func loadImage(from url: URL) {
@@ -193,6 +102,17 @@ class ViewController: UIViewController {
                 self.showErrorAlert(message: "Failed to load image.")
             }
         }
+    }
+    
+    @IBAction func safariButtonTapped(_ sender: UIButton) {
+        guard currentImageIndex >= 0, currentImageIndex < imageHistory.count else {
+            print("No valid image URL available")
+            return
+        }
+        
+        let currentURL = imageHistory[currentImageIndex]
+        let safariVC = SFSafariViewController(url: currentURL)
+        present(safariVC, animated: true, completion: nil)
     }
     
     @IBAction func refreshButtonTapped(_ sender: UIButton) {
